:hammer: Rework MC BatchNorm
